requirements.txt
Flask==3.0.3
gunicorn==22.0.0
psycopg[binary]==3.2.9
python-dotenv==1.0.1
